+++ quam/components/quantum_components/quantum_component.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 from dataclasses import field
 from typing import Any, Dict, Union
 from quam.core.quam_classes import quam_dataclass, QuamComponent
-from quam.components.operations import BaseOperation
+from quam.components.implementations import BaseImplementation
 
 __all__ = ["QuantumComponent"]
 
@@ -10,7 +10,7 @@ __all__ = ["QuantumComponent"]
 @quam_dataclass
 class QuantumComponent(QuamComponent, ABC):
     id: Union[str, int]
-    operations: Dict[str, BaseOperation] = field(default_factory=dict)
+    operations: Dict[str, BaseImplementation] = field(default_factory=dict)
 
     @property
     @abstractmethod
