+++ quam/examples/superconducting_qubits/implementations.py
@@ -1,44 +1,13 @@
-from abc import ABC
 from copy import copy
 
-from quam.components.pulses import Pulse
-from quam.components.operations import BaseOperation
-from quam.components.quantum_components.qubit import Qubit
 from quam.core import quam_dataclass
+from quam.components.pulses import Pulse
+from quam.components.implementations import QubitPairImplementation
 from quam.utils import string_reference as str_ref
 
 
-__all__ = ["QubitPairOperation", "CZOperation"]
-
-
 @quam_dataclass
-class QubitPairOperation(BaseOperation, ABC):
-    @property
-    def qubit_pair(self):  # TODO Add QubitPair return type
-        from quam.components.quantum_components.qubit_pair import QubitPair
-
-        if isinstance(self.parent, QubitPair):
-            return self.parent
-        elif hasattr(self.parent, "parent") and isinstance(
-            self.parent.parent, QubitPair
-        ):
-            return self.parent.parent
-        else:
-            raise AttributeError(
-                "TwoQubitGate is not attached to a QubitPair. 2Q_gate: {self}"
-            )
-
-    @property
-    def qubit_control(self) -> Qubit:
-        return self.qubit_pair.qubit_control
-
-    @property
-    def qubit_target(self) -> Qubit:
-        return self.qubit_pair.qubit_target
-
-
-@quam_dataclass
-class CZOperation(QubitPairOperation):
+class CZImplementation(QubitPairImplementation):
     """CZ Operation for a qubit pair"""
 
     flux_pulse_control: Pulse
